bestsinglestackwithqb moved to own, needs p/d stil
