+++ dfs_positions_calc_funcs/stackingValuesCalcs.js
@@ -8,15 +8,16 @@ const allWRs = require('../teamandpositionvariables/allWRVariables');
 const allTEs = require('../teamandpositionvariables/allTEVariables');
 
 const QBWROneSingleStackData = require('./stackingValuesCalcs/bestQBWROneStacks');
+const bestTeamSingleStackWithQB = require('./stackingValuesCalcs/bestSingleStackWithQB');
 
-const allTeamsBestHalfSingleStackWithQB = [];
-const allTeamsBestHalfSingleStackWithQBWithNames = [];
+// const allTeamsBestHalfSingleStackWithQB = [];
+// const allTeamsBestHalfSingleStackWithQBWithNames = [];
 
-const allTeamsBestPPRSingleStackWithQB = [];
-const allTeamsBestPPRSingleStackWithQBWithNames = [];
+// const allTeamsBestPPRSingleStackWithQB = [];
+// const allTeamsBestPPRSingleStackWithQBWithNames = [];
 
-const allTEPremiumBestSingleStackWithQB = [];
-const allTEPremiumBestSingleStackWithQBWithNames = [];
+// const allTEPremiumBestSingleStackWithQB = [];
+// const allTEPremiumBestSingleStackWithQBWithNames = [];
 
 const allTeamsBestHalfDoubleStackWithQBAndWROne = [];
 const allTeamsBestHalfDoubleStackWithQBAndWROneWithNames = [];
@@ -45,67 +46,67 @@ const allPPRFlexValuesWithNames = [];
 const allTEPFlexValuesWithNames = [];
 
 const allStackingCalcFuncs = {
-  CalcTeamBestSingleStackWithQB(array, i) {
-    // console.log(array);
-    let [qb, rbOne, rbTwo, wrOne, wrTwo, te] = array;
+  // CalcTeamBestSingleStackWithQB(array, i) {
+  //   // console.log(array);
+  //   let [qb, rbOne, rbTwo, wrOne, wrTwo, te] = array;
 
-    let qbrbOne = qb + rbOne;
-    let qbrbTwo = qb + rbTwo;
-    let qbwrOne = qb + wrOne;
-    let qbwrTwo = qb + wrTwo;
-    let qbte = qb + te;
+  //   let qbrbOne = qb + rbOne;
+  //   let qbrbTwo = qb + rbTwo;
+  //   let qbwrOne = qb + wrOne;
+  //   let qbwrTwo = qb + wrTwo;
+  //   let qbte = qb + te;
 
-    let qbrbOneWithNames = `${qbrbOne}: QB-RB1- ${allQBs[i].name}-${allRBs[i].RBOne.name}`;
-    let qbrbTwoWithNames = `${qbrbTwo}: QB-RB2- ${allQBs[i].name}-${allRBs[i].RBTwo.name}`;
-    let qbwrOneWithNames = `${qbwrOne}: QB-WR1- ${allQBs[i].name}-${allWRs[i].WROne.name}`;
-    let qbwrTwoWithNames = `${qbwrTwo}: QB-WR2- ${allQBs[i].name}-${allWRs[i].WRTwo.name}`;
-    let qbteWithNames = `${+qbte.toFixed(2)}: QB-TE- ${allQBs[i].name}-${
-      allTEs[i].TE.name
-    }`;
+  //   let qbrbOneWithNames = `${qbrbOne}: QB-RB1- ${allQBs[i].name}-${allRBs[i].RBOne.name}`;
+  //   let qbrbTwoWithNames = `${qbrbTwo}: QB-RB2- ${allQBs[i].name}-${allRBs[i].RBTwo.name}`;
+  //   let qbwrOneWithNames = `${qbwrOne}: QB-WR1- ${allQBs[i].name}-${allWRs[i].WROne.name}`;
+  //   let qbwrTwoWithNames = `${qbwrTwo}: QB-WR2- ${allQBs[i].name}-${allWRs[i].WRTwo.name}`;
+  //   let qbteWithNames = `${+qbte.toFixed(2)}: QB-TE- ${allQBs[i].name}-${
+  //     allTEs[i].TE.name
+  //   }`;
 
-    let bestSingleStackWithQB = qbrbOne;
-    let bestSingleStackWithQBAndNames = qbrbOneWithNames;
+  //   let bestSingleStackWithQB = qbrbOne;
+  //   let bestSingleStackWithQBAndNames = qbrbOneWithNames;
 
-    //best single stack
-    if (bestSingleStackWithQB < qbrbTwo) {
-      bestSingleStackWithQB = qbrbTwo;
-      bestSingleStackWithQBAndNames = qbrbTwoWithNames;
-    }
-    if (bestSingleStackWithQB < qbwrOne) {
-      bestSingleStackWithQB = qbwrOne;
-      bestSingleStackWithQBAndNames = qbwrOneWithNames;
-    }
-    if (bestSingleStackWithQB < qbwrTwo) {
-      bestSingleStackWithQB = qbwrTwo;
-      bestSingleStackWithQBAndNames = qbwrTwoWithNames;
-    }
-    if (bestSingleStackWithQB < qbte) {
-      bestSingleStackWithQB = qbte;
-      bestSingleStackWithQBAndNames = qbteWithNames;
-    }
+  //   //best single stack
+  //   if (bestSingleStackWithQB < qbrbTwo) {
+  //     bestSingleStackWithQB = qbrbTwo;
+  //     bestSingleStackWithQBAndNames = qbrbTwoWithNames;
+  //   }
+  //   if (bestSingleStackWithQB < qbwrOne) {
+  //     bestSingleStackWithQB = qbwrOne;
+  //     bestSingleStackWithQBAndNames = qbwrOneWithNames;
+  //   }
+  //   if (bestSingleStackWithQB < qbwrTwo) {
+  //     bestSingleStackWithQB = qbwrTwo;
+  //     bestSingleStackWithQBAndNames = qbwrTwoWithNames;
+  //   }
+  //   if (bestSingleStackWithQB < qbte) {
+  //     bestSingleStackWithQB = qbte;
+  //     bestSingleStackWithQBAndNames = qbteWithNames;
+  //   }
 
-    if (rbOne === allRBData.allRBOneHalfPPRProjectedPointsValues[i]) {
-      allTeamsBestHalfSingleStackWithQB.push(bestSingleStackWithQB);
-      allTeamsBestHalfSingleStackWithQBWithNames.push(
-        bestSingleStackWithQBAndNames
-      );
-    }
-    if (
-      rbOne === allRBData.allRBOneFullPPRProjectedPointsValues[i] &&
-      te === allTEData.allFullTEFinalProjectedPointsValues[i]
-    ) {
-      allTeamsBestPPRSingleStackWithQB.push(bestSingleStackWithQB);
-      allTeamsBestPPRSingleStackWithQBWithNames.push(
-        bestSingleStackWithQBAndNames
-      );
-    }
-    if (te === allTEData.allTEPremiuimFinalProjectedPointsValues[i]) {
-      allTEPremiumBestSingleStackWithQB.push(bestSingleStackWithQB);
-      allTEPremiumBestSingleStackWithQBWithNames.push(
-        bestSingleStackWithQBAndNames
-      );
-    }
-  },
+  //   if (rbOne === allRBData.allRBOneHalfPPRProjectedPointsValues[i]) {
+  //     allTeamsBestHalfSingleStackWithQB.push(bestSingleStackWithQB);
+  //     allTeamsBestHalfSingleStackWithQBWithNames.push(
+  //       bestSingleStackWithQBAndNames
+  //     );
+  //   }
+  //   if (
+  //     rbOne === allRBData.allRBOneFullPPRProjectedPointsValues[i] &&
+  //     te === allTEData.allFullTEFinalProjectedPointsValues[i]
+  //   ) {
+  //     allTeamsBestPPRSingleStackWithQB.push(bestSingleStackWithQB);
+  //     allTeamsBestPPRSingleStackWithQBWithNames.push(
+  //       bestSingleStackWithQBAndNames
+  //     );
+  //   }
+  //   if (te === allTEData.allTEPremiuimFinalProjectedPointsValues[i]) {
+  //     allTEPremiumBestSingleStackWithQB.push(bestSingleStackWithQB);
+  //     allTEPremiumBestSingleStackWithQBWithNames.push(
+  //       bestSingleStackWithQBAndNames
+  //     );
+  //   }
+  // },
 
   CalcTeamBestDoubleAndTripleStacksWithQBAndQBsWROne(array, i) {
     let [
@@ -689,16 +690,16 @@ allQBData.allQBFinalProjectedPointsValues.map(function (value, i) {
   let tePPR = allTEData.allFullTEFinalProjectedPointsValues[i];
   let teTEPremium = allTEData.allTEPremiuimFinalProjectedPointsValues[i];
 
-  let halfArray = [qb, rbOneHalf, rbTwoHalf, wrOneHalf, wrTwoHalf, teHalf];
-  let fullArray = [qb, rbOnePPR, rbTwoPPR, wrOnePPR, wrTwoPPR, tePPR];
-  let TEPremiumArray = [
-    qb,
-    rbOnePPR,
-    rbTwoPPR,
-    wrOnePPR,
-    wrTwoPPR,
-    teTEPremium,
-  ];
+  // let halfArray = [qb, rbOneHalf, rbTwoHalf, wrOneHalf, wrTwoHalf, teHalf];
+  // let fullArray = [qb, rbOnePPR, rbTwoPPR, wrOnePPR, wrTwoPPR, tePPR];
+  // let TEPremiumArray = [
+  //   qb,
+  //   rbOnePPR,
+  //   rbTwoPPR,
+  //   wrOnePPR,
+  //   wrTwoPPR,
+  //   teTEPremium,
+  // ];
 
   if (i % 2 === 1) {
     opposingRBOneHalf = allRBData.allRBOneHalfPPRProjectedPointsValues[i - 1];
@@ -799,14 +800,14 @@ allQBData.allQBFinalProjectedPointsValues.map(function (value, i) {
     oppteName,
   ];
 
-  const bestHalfTeamSingleStackWithQB =
-    allStackingCalcFuncs.CalcTeamBestSingleStackWithQB(halfArray, i);
+  // const bestHalfTeamSingleStackWithQB =
+  //   allStackingCalcFuncs.CalcTeamBestSingleStackWithQB(halfArray, i);
 
-  const bestPPRTeamSingleStackWithQB =
-    allStackingCalcFuncs.CalcTeamBestSingleStackWithQB(fullArray, i);
+  // const bestPPRTeamSingleStackWithQB =
+  //   allStackingCalcFuncs.CalcTeamBestSingleStackWithQB(fullArray, i);
 
-  const bestTEPremiumTeamSingleStackWithQB =
-    allStackingCalcFuncs.CalcTeamBestSingleStackWithQB(TEPremiumArray, i);
+  // const bestTEPremiumTeamSingleStackWithQB =
+  //   allStackingCalcFuncs.CalcTeamBestSingleStackWithQB(TEPremiumArray, i);
 
   const bestHalfDoubleStackWithQBAndQBsWROne =
     allStackingCalcFuncs.CalcTeamBestDoubleAndTripleStacksWithQBAndQBsWROne(
@@ -1114,16 +1115,19 @@ const allStackData = {
     QBWROneSingleStackData.allHalfQBWROneStackWithNames,
   allPPRQBWROneStacksWithNames:
     QBWROneSingleStackData.allPPRQBWROneStacksWithNames,
-  allTeamsBestHalfSingleStackWithQB: allTeamsBestHalfSingleStackWithQB,
+  allTeamsBestHalfSingleStackWithQB:
+    bestTeamSingleStackWithQB.allTeamsBestHalfSingleStackWithQB,
   allTeamsBestHalfSingleStackWithQBWithNames:
-    allTeamsBestHalfSingleStackWithQBWithNames,
+    bestTeamSingleStackWithQB.allTeamsBestHalfSingleStackWithQBWithNames,
   // sortedAllHalfQBWROneStacks: sortedAllHalfQBWROneStacks,
-  allTeamsBestPPRSingleStackWithQB: allTeamsBestPPRSingleStackWithQB,
+  allTeamsBestPPRSingleStackWithQB:
+    bestTeamSingleStackWithQB.allTeamsBestPPRSingleStackWithQB,
   allTeamsBestPPRSingleStackWithQBWithNames:
-    allTeamsBestPPRSingleStackWithQBWithNames,
-  allTEPremiumBestSingleStackWithQB: allTEPremiumBestSingleStackWithQB,
+    bestTeamSingleStackWithQB.allTeamsBestPPRSingleStackWithQBWithNames,
+  allTEPremiumBestSingleStackWithQB:
+    bestTeamSingleStackWithQB.allTEPremiumBestSingleStackWithQB,
   allTEPremiumBestSingleStackWithQBWithNames:
-    allTEPremiumBestSingleStackWithQBWithNames,
+    bestTeamSingleStackWithQB.allTEPremiumBestSingleStackWithQBWithNames,
 
   halfDKBestQBWROneStackspointsperdollar:
     QBWROneSingleStackData.halfDKBestQBWROneStackspointsperdollar,
@@ -1165,7 +1169,7 @@ const allStackData = {
     allTEPremiumBestTripleStackWithQBAndWROneWithNames,
 };
 
-// console.log(allStackData);
+console.log(allStackData);
 
 module.exports = allStackingCalcFuncs;
 module.exports = allStackData;
